+++ src/main/java/duke/TasksType/Deadline.java
@@ -1,7 +1,6 @@
 package duke.TasksType;
 
 import duke.Utility.Util;
-
 import java.time.LocalDateTime;
 
 public class Deadline extends Task
