+++ src/test/java/duke/TaskListTest.java
@@ -25,4 +25,21 @@ public class TaskListTest {
 
         assertEquals(Task.priorityLevel.High, taskList.getTask(0).getPriorityLevel());
     }
+    /**
+     * This JUnit tests the TaskList class, addTask() method
+     * adds two tasks and asserts that the description is stored and retrieved correctly
+     */
+    @Test
+    public void addTask() {
+        ArrayList<Task> list = new ArrayList<>();
+        Task task = new Todo("Eat dinner");
+        list.add(task);
+        Task task2 = new Todo("Eat breakfast");
+        list.add(task2);
+        TaskList taskList = new TaskList();
+        taskList.storeList(list);
+
+        assertEquals("Eat dinner", taskList.getTask(0).getDescription());
+        assertEquals("Eat breakfast", taskList.getTask(1).getDescription());
+    }
 }
